fix/routing: avoid erroring if elder is no longer in ElderInfo
When introducing demotion, elder may find they are no longer in
ElderInfo. The problem is we get an error because the node is not one
of the participants and so did not store the key.

Store the keys always, and use the correct one accumulating a new
SectionInfo.

Test:
clippy + soak test