fix/routing: differentiate Event for relocation
With this test:
SEED="[2938108988, 210465919, 1400964886, 1323843700]" cargo test --release --features=mock -- messages_during_churn

Relocation occurs during the node creation to reach split.
This may be an issue in itself, but the problem is that the Connected
event occurs again and the test do not expect it.

Ignore Connected event as a result of relocation.