fix/routing: Remove unneeded ConnectionRequest handling
ConnectionRequest is now mostly unused, only to try to re-establish
a connection with a lost elder. This is the remaining part of the work
to use responsiveness as an alternative to having a live connection.

However, this part of the code is no longer needed at this stage as
there is no need to try to find the connection information since we
already have them.

Test:
clippy + soak test