fix/routing: process signature accumulation as an adult
An Adult may receive Signature messages to accumulate if it does not
yet know it has been promoted, or if it has been demoted.

In both case, we should not wait to be an adult to process them. We
can accumulate and forward the signed routing messages to all our elders
for sending further. We can optimize this later prossibly only sending
to 3 of our elders.

Test:
clippy + soak test.