refactor/elder: consistent reset pre-process
Ensure that all nodes process the reset in the same way (i.e demoted
node during split should process it the same way as other nodes).

Also ensure we have all the information to allow eventually sending
a message if we have Online vote and we need to notify the node to send
more JoinRequest.

Test:
clippy + soak.