fix/chain: ensure churn is backloged when relocation is in progress
Fixes:
SEED="[4205599700, 2193644365, 693838133, 1971124]" RUST_LOG=trace cargo test --release --features=mock -- aggressive_churn

When a relocation is in progress we should not be adding or removing
nodes that may change the section. Because the vote is signed, we are
not re-voting for the relocate event when changing elders.

Test:
Verify seed pass.
clippy + soak tests