fix/routing: handle polling relocation like normal polling
Currently polling for relocation is tightly coupled with Online/Offline/
SectionInfo events. This lead to duplication and also issues.

In particular, the poll_relocation call in handle_section_info_event
was breaking shared state since elder seing this could have
relocation_in_progress=true, while Adults only getting the genesis info
would not. The current approach does not have this issue and so we do
not have to move that state to SharedState.

Test:
clippy + soak tests