fix/elder: store unaccumulated info as SignedMessage and always resend
We want to be able to extend that functionality to other kind of
messages.

If we have any unconsensused observation that is still valid we should
process it. It is a rare occurence since it happens on Elder churn/split

PendingMessageKey should be enough information to be unique,
so for example if get 2 NeighbourInfo successively we keep both to
resend as needed, as both need to reach consensus and accumulation.

Test:
clippy + soak test