fix/adult: avoid bootstrap loop when Adult not know they are Elders
This is not necessarily an ideal way of doing things, however,
there are a number of item in the relocation/join/bootstrap that we
need to re-visit, in particular in the realy of security.

For example a bad Elder may send a relocating node to a fake network
where it will rebootstrap forever (or enough to clear its bootstrap
cache).

Failing seed:
SEED="[2158704635, 314330066, 3143740410, 2206775824]" cargo test --release --features=mock -- messages_during_churn

Test:
Verify the seed is fixed.