fix/adult: do not disconnect from relocated node
Adults were disconnecting from relocated node, elder did not.
The problem in this seed is that the Adult will do that as it catches
up with Parsec, as it is about to be promoted. So we don't want to
disconnect from a potential neighbour.

Fixes:
SEED="[3467112354, 1436145184, 989989653, 3612989132]" RUST_LOG=trace cargo test --release --features=mock -- aggressive_churn

Test:
clippy + soak test