fix/chain: only combine valid sigs for accumulated events
We need to create a valid signature. With an accumulated event, we will
always have threshold + 1 valid signature share. We may have invalid
shares as well that we need to ignore when combining signature to get
a valid signature.

Clean up chain test and add new test to cover functionality for
Relocate.

Note:
Use log or pannic since if the event reach quorum, then enough valid
node should have signed it that it is not possible for a malicious node
to make the signature combination fail.

Test:
New unit tests.
clippy + soak test