routing/refactor Authority no longer valid
Renamed Authority to Location. This is the next phase of cleanup and does appear to make more
locial sense when reading the code. It is likely Authority will make a come back though, but as
an actual authority that we can query for validity. Desing considerations though on whether
Authority will be src dst or content. There is reasonable argument that Authority should be for
whole message.

routing/docs

Updated documentation for Location struct.

routing/fixup

Rename authorities -> locations in docs.