fix/routing: re-use serialized hop messages
When forwarding hop messages, we use to de-serialize and then
serialize again that content again to be able to re-send it.

Not only do we use CPU for the re-serialization, we also use more memory
as the message Bytes we receive can be used directly to resend using
quic_p2p.

This also open new improvments:
 - Filter more effectively as we already have serialized data.
 - Avoid Deserializing the full message if possible doing partial
   deserialization to extract the info we need.

Test:
clippy + test