fix/routing: avoid serializing for filter and include signature
We use the final message serialization for filtering messages so we do
not have to serailize again hop messages.

This also address an attack where a message with wrong signature would
lead to follow up messages with correct signature being ignored:
 - We now include signature in filtering so this is no longer possible.

Note: This also open the door to filtering earlier, before we even
deserialize.

Test:
Clippy + test