Don't fail/panic when trust check fails on too new proof
When we are elder and we receive a message whose proof is too new, that can only mean one thing (assuming non-faulty sender): our section has already demoted us, we just haven't caught up to it yet.

This is because in order for honest section to send us a message proven with version X, our section must have previously sent them `AckMessage` with version X. But `AckMessage` requires total consensus (all elders must vote), but we haven't voted so the only logical conclusion is that we've been demoted.

This means that the correct thing to do is the same thing that we would do as adult - forward the message to our elders.