feat/routing: partial message deserialization for hops
Most messages will require little deserializations:
 - Either they'll be filtered as hop message for duplicate
 - Either they will not be the final destination and only require dst.

We serialize/deserialize Message as tupple, and for hop message we
deserialize only the first field `dst` that is needed to forward messages.
When we need the full message to process it, we then take it (if we
had it already, or deserialize it).

Test:
clippy + tests