Remove message_sent test hack
Explanation of the hack, why it was needed and why it's no longer the case:

To be able to tell when to stop polling the test network, we have the concept of node being busy. The polling then stops only if no node is busy. Node is considered busy if it handled any network Event or Action in the current iteration (there are other conditions but they are not relevant for this explanation). Problem is that we have some actions that are triggered periodically and so those would always make a node seem busy and the polling would never stop. Enter hack #1, which is that we don't consider a node as busy if it only handled a Timeout action.

However, a node might do some actual work in reaction to a Timeout and so this would make an actual busy node seem not-busy. Enter hack #2, which is we additionally consider a node busy if it sent any message during the current iteration.

That however clashes with parsec gossip because it is sent periodically and unconditionally and so we would again consider node as busy if it only sent a parsec gossip and the polling would never stop. Enter hack #3, which is to ignore parsec gossip messages when determining whether a message was sent.

This was necessary for some time but turns out the only thing we now do in reaction to a Timeout is to send parsec gossip. So hack #2 and hack #3 cancel each other out and are therefore not necessary. So this commit removes them.