Remove elder connection check
This check was supposed to detect the situation where we lose connections to all our elders and trigger a node restart in that case. The check require peer map which we are going to remove, so it would no longer be possible to perform it. It also had false positives, because even when we get disconnected from all elders, we still have their connection details and can re-contact them.

This check might possibly be brought back in the future where it would likely be based on checking responsiveness instead of physical connection.