Limit the number of simultaneously dropped nodes in the churn tests
The limit is set as the difference between `safe_section_size` and `elder_size` per section. This should guarantee that a section never drops below `elder_size` nodes even in case of split. This criterion is stronger than the ones we had there before which allowed us to also simplify the code.