test: do not panic if no gossip recipient is available
It can sometimes happen that none of the available gossip recipients have known network addresses. This can happen if all the known ones are blocked by the gossip limiter and the only remaining ones are still in the process of joining.

More detailed explanation: it can happen that when a new node is joining, the `StartDkg` event reaches consensus in `Parsec` before the `Online` event gets accumulated in `Chain`, even though the `Online` event reached parsec consensus first. The `StartDkg` then modifies the list of gossip recipients so such node might get selected as the next gossip recipient. But as we haven't yet accumulated the `Online`, we don't know their network address and thus cannot send the gossip to them.