Ignore ConnectionFailure as BootstrappingPeer
Consider bootstrap request as failed only when it timeouts. This fixes the following situation:

A node gets relocated. It transitions to `BootstrappingPeer` and sends `BootstrapRequest`s. The recipients of the `BootstrapRequest`s are  former neighbours of the node. They accumulate `NeighbourInfo` realizing that the node is no longer their neighbour and disconnect from it. Only after that they receive the `BootstrapRequest` which they respond to normally. The node then receives `ConnectionFailure` and marks the peer as failed which causes it to ignore any subsequent messages (including `BootstrapResponse`) from it.

This change is in line with the recent changes to not treat disconnected peers as lost.