Add section version and use it when deciding promotions
This commit adds `section_version` to `SharedState` which is a number that increments every time a node joins or leaves the section. Every member gets assigned the version at the time it joined the section. This is then used to break ties when deciding which node to promote in case there is more than one with the same age counter - those with smaller section version (that is, those that joined earlier) are preferred.