Do not clear already assigned relocation overrides
So when a actual relocation destination X is overridden with a different destination Y, such override will stay in effect forever (that is, until the end of the current test) even when the `RelocationOverride` object goes out of scope. Only thing that gets cleared is the override of any future relocations.

The reason for this change is to avoid consensus breakage when some nodes already processed a relocation but others have not at the time when the `RelocationOverride` object goes out of scope.