test: remove assert that a non-elder can't have any neighbour info
A node might have been already promoted and sent the initial shared state but still not caught up to its own promotion. In this case such node would see itself as non-elder but might already have some of the info that normally only elders have. This is not an error.