test: allow exhausting max iterations in the poll_all function
That function does not advance time but sometimes time advancement is necessary to make progress. This change makes it so that exhausting the max iterations is not an error to allow the outer loop (in `poll_until`) to advance the time and try again. This does not prevent any genuine errors to be caught because the outer loop still has hard iteration limit.