handle Relocate message also as elder
When relocating an elder, we first demote them to adult, then wait for a "relocation cool-down period" and only then send them the `Relocate` message. This makes it very unlikely that they would still see themselves as elder by the time they receive it. It's not impossible however and when it happens, the node must still proceed with the relocation because it must always obey section decisions. However, we weren't doing it and so this commit fixes it.