discard more unexpected message types in the Bootstrapping stage
The only message types that are still bounced are `NeighbourInfo` and `UserMessage`. The idea is that we can only receive those messages if we are being relocated and they were sent while we were still a member of our old section. Bouncing them will cause them to be resent later at which point we might be already joined in the new section and will be able to route them to their intended destination.