allow keeping compatible prefixes in the PrefixMap in some case
This is in preparation to have the entries in the maps signed. The signature would sign both the key (the prefix) and the value which means we can no longer move an entry from its prefix to its child prefix when the sibling is inserted, because that would invalidate the signature. Because of this we now allow a prefix and its descendant(s) to exist simultaneously in the map as long as the descendants don't fully cover it. Once they do, the prefix is removed.