Make Message immutable.
Remove MessageWithBytes and PartialMessage. Now message serialialises at most once and calculates
signature once as well as calculate hash once. These were memoized in previous iterations but deemed
to be always useful as we must sign message, we must confirm sig on receipt. We deserialise from
the wire as bytes and keep that in case it's required (to forward or bounce the message). We filter
via hash too (except Direct messages) therefor it made snse to not complexify via memoization but
instead keep the serialised and hash fields correct.

The message type now has unfailable to_bytes() and hash() methods which does help with cleaner code.