add MemberState::Relocated
This is used instead of `MemberState::Left` when the node leaves due to relocation. This is needed to cover the following edge case:

Some elders vote to relocate node A to destination X and others vote to relocate it to Y. Neither is enough for accumulation, but both also vote the node Offline and because both Offlines are identical, they do accumulate. By using a separate `Relocated` state which contain the relocation destination, we make the votes distinct and therefore not accumulate either.