fix: prevent losing incoming messages during bootstrapping
The problem was during bootstrap, if we received the `NodeApproval` message followed immediately by some other message (say `Relocate`), then the `NodeApproval` would be processed and the bootstrap would complete, leaving the following message in the intermediary channel buffer never to be taken out of it and thus losing it.

The fix is to remove the intermediary channel and replace it with a simple wrapper around the `ConnectionEvent` receiver. Thus changing a the "push/pull" model into a simple "pull" model. This way, a message is never retrieved from the channel if we are not ready to process it.