refactor!: remove `Error::UntrustedMessage`
Replaced with `InvalidMessage` which is more appropriate.

`UntrustedMessage` was supposed to mean that the message is otherwise valid but we don't trust the key it was signed with. However, we often used this error even in cases where the message was actually invalid, for example when the proof chain was broken. Also, untrusted messages are bounced back to the sender, so receiving them is not treated as an error anyway.

BREAKING CHANGE: this affects the `Error` type which is a part of the public API.